requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Pillow==10.1.0
google-generativeai==0.3.2
requests==2.31.0
Werkzeug==3.0.0
WTForms==3.1.1
SQLAlchemy==2.0.23
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-barcode==0.15.1
qrcode==7.4.2
